requirements.txt
absl-py==2.1.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.0
cycler==0.12.1
datasets==3.3.1
dill==0.3.8
distro==1.9.0
dm-tree==0.1.8
evaluate==0.4.3
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.17.0
fire==0.7.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==25.2.10
fonttools==4.55.0
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2024.12.0
ftfy==6.3.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.0
icon-font-to-png==0.4.1
idna==3.4
importlib-metadata==6.8.0
importlib_resources==6.4.5
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.9.0
joblib==1.3.2
kagglehub==0.3.8
keras==3.8.0
keras-core==0.1.7
keras-hub==0.18.1
keras-nlp==0.18.1
kiwisolver==1.4.7
libclang==18.1.1
lxml==4.9.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
multitasking==0.0.11
namex==0.0.8
networkx==3.2.1
numpy==1.26.1
openai==1.70.0
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
palettable==3.3.3
pandas==2.1.1
peewee==3.17.0
Pillow==8.4.0
protobuf==4.25.6
pyaml==24.9.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyodbc==5.2.0
pyparsing==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.3.2
scikit-optimize==0.10.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.46.1
stylecloud==0.5.2
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-text==2.16.1
termcolor==2.5.0
threadpoolctl==3.2.0
tinycss==0.4
tokenizers==0.21.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
Werkzeug==3.0.0
wordcloud==1.9.4
wrapt==1.17.2
xgboost==2.0.0
xxhash==3.5.0
yarl==1.9.2
yellowbrick==1.5
yfinance==0.2.31
zipp==3.17.0
